src/python/requirements.txt
tensorflow
numpy
Pillow
opencv-python
seaborn
datetime
email-to
